Refactor(#30-battery-view): Edit 수면량 목표 화면 파일로 분리
